main/src/requirements.txt
celery==5.3.6
Django==5.0.2
djangorestframework==3.14.0
redis==5.0.1
requests==2.31.0
psycopg2-binary==2.9.1
django-environ==0.11.2
